настройка для 64-bit окружения

+++ src/util.d
@@ -7,7 +7,7 @@ string fromStringz(char* s)
         return cast(string) (s ? s[0 .. strlen(s)] : null);
 }
 
-string fromStringz(char* s, int len)
+string fromStringz(char* s, ulong len)
 {
         return cast(string) (s ? s[0 .. len] : null);
 }
